# huff0: Improve 4X decompression speed 5-10%