--- a/huff0/decompress.go
+++ b/huff0/decompress.go
@@ -20,9 +20,8 @@ type dEntrySingle struct {
 
 // double-symbols decoding
 type dEntryDouble struct {
-	seq   uint16
+	seq   [2]byte
 	nBits uint8
-	len   uint8
 }
 
 // Uses special code for all tables that are < 8 bits.
@@ -914,7 +913,7 @@ func (d *Decoder) decompress4X8bit(dst, src []byte) ([]byte, error) {
 	out := dst
 	dstEvery := (dstSize + 3) / 4
 
-	shift := (8 - d.actualTableLog) & 7
+	shift := (56 + (8 - d.actualTableLog)) & 63
 
 	const tlSize = 1 << 8
 	single := d.dt.single[:tlSize]
@@ -938,37 +937,41 @@ func (d *Decoder) decompress4X8bit(dst, src []byte) ([]byte, error) {
 			br[stream].fillFast()
 			br[stream2].fillFast()
 
-			v := single[br[stream].peekByteFast()>>shift].entry
+			v := single[uint8(br[stream].value>>shift)].entry
+			v2 := single[uint8(br[stream2].value>>shift)].entry
 			buf[off+bufoff*stream] = uint8(v >> 8)
-			br[stream].advance(uint8(v))
-
-			v2 := single[br[stream2].peekByteFast()>>shift].entry
 			buf[off+bufoff*stream2] = uint8(v2 >> 8)
-			br[stream2].advance(uint8(v2))
+			br[stream].bitsRead += uint8(v)
+			br[stream].value <<= v & 63
+			br[stream2].bitsRead += uint8(v2)
+			br[stream2].value <<= v2 & 63
 
-			v = single[br[stream].peekByteFast()>>shift].entry
+			v = single[uint8(br[stream].value>>shift)].entry
+			v2 = single[uint8(br[stream2].value>>shift)].entry
 			buf[off+bufoff*stream+1] = uint8(v >> 8)
-			br[stream].advance(uint8(v))
-
-			v2 = single[br[stream2].peekByteFast()>>shift].entry
 			buf[off+bufoff*stream2+1] = uint8(v2 >> 8)
-			br[stream2].advance(uint8(v2))
+			br[stream].bitsRead += uint8(v)
+			br[stream].value <<= v & 63
+			br[stream2].bitsRead += uint8(v2)
+			br[stream2].value <<= v2 & 63
 
-			v = single[br[stream].peekByteFast()>>shift].entry
+			v = single[uint8(br[stream].value>>shift)].entry
+			v2 = single[uint8(br[stream2].value>>shift)].entry
 			buf[off+bufoff*stream+2] = uint8(v >> 8)
-			br[stream].advance(uint8(v))
-
-			v2 = single[br[stream2].peekByteFast()>>shift].entry
 			buf[off+bufoff*stream2+2] = uint8(v2 >> 8)
-			br[stream2].advance(uint8(v2))
+			br[stream].bitsRead += uint8(v)
+			br[stream].value <<= v & 63
+			br[stream2].bitsRead += uint8(v2)
+			br[stream2].value <<= v2 & 63
 
-			v = single[br[stream].peekByteFast()>>shift].entry
+			v = single[uint8(br[stream].value>>shift)].entry
+			v2 = single[uint8(br[stream2].value>>shift)].entry
 			buf[off+bufoff*stream+3] = uint8(v >> 8)
-			br[stream].advance(uint8(v))
-
-			v2 = single[br[stream2].peekByteFast()>>shift].entry
 			buf[off+bufoff*stream2+3] = uint8(v2 >> 8)
-			br[stream2].advance(uint8(v2))
+			br[stream].bitsRead += uint8(v)
+			br[stream].value <<= v & 63
+			br[stream2].bitsRead += uint8(v2)
+			br[stream2].value <<= v2 & 63
 		}
 
 		{
@@ -977,37 +980,41 @@ func (d *Decoder) decompress4X8bit(dst, src []byte) ([]byte, error) {
 			br[stream].fillFast()
 			br[stream2].fillFast()
 
-			v := single[br[stream].peekByteFast()>>shift].entry
+			v := single[uint8(br[stream].value>>shift)].entry
+			v2 := single[uint8(br[stream2].value>>shift)].entry
 			buf[off+bufoff*stream] = uint8(v >> 8)
-			br[stream].advance(uint8(v))
-
-			v2 := single[br[stream2].peekByteFast()>>shift].entry
 			buf[off+bufoff*stream2] = uint8(v2 >> 8)
-			br[stream2].advance(uint8(v2))
+			br[stream].bitsRead += uint8(v)
+			br[stream].value <<= v & 63
+			br[stream2].bitsRead += uint8(v2)
+			br[stream2].value <<= v2 & 63
 
-			v = single[br[stream].peekByteFast()>>shift].entry
+			v = single[uint8(br[stream].value>>shift)].entry
+			v2 = single[uint8(br[stream2].value>>shift)].entry
 			buf[off+bufoff*stream+1] = uint8(v >> 8)
-			br[stream].advance(uint8(v))
-
-			v2 = single[br[stream2].peekByteFast()>>shift].entry
 			buf[off+bufoff*stream2+1] = uint8(v2 >> 8)
-			br[stream2].advance(uint8(v2))
+			br[stream].bitsRead += uint8(v)
+			br[stream].value <<= v & 63
+			br[stream2].bitsRead += uint8(v2)
+			br[stream2].value <<= v2 & 63
 
-			v = single[br[stream].peekByteFast()>>shift].entry
+			v = single[uint8(br[stream].value>>shift)].entry
+			v2 = single[uint8(br[stream2].value>>shift)].entry
 			buf[off+bufoff*stream+2] = uint8(v >> 8)
-			br[stream].advance(uint8(v))
-
-			v2 = single[br[stream2].peekByteFast()>>shift].entry
 			buf[off+bufoff*stream2+2] = uint8(v2 >> 8)
-			br[stream2].advance(uint8(v2))
+			br[stream].bitsRead += uint8(v)
+			br[stream].value <<= v & 63
+			br[stream2].bitsRead += uint8(v2)
+			br[stream2].value <<= v2 & 63
 
-			v = single[br[stream].peekByteFast()>>shift].entry
+			v = single[uint8(br[stream].value>>shift)].entry
+			v2 = single[uint8(br[stream2].value>>shift)].entry
 			buf[off+bufoff*stream+3] = uint8(v >> 8)
-			br[stream].advance(uint8(v))
-
-			v2 = single[br[stream2].peekByteFast()>>shift].entry
 			buf[off+bufoff*stream2+3] = uint8(v2 >> 8)
-			br[stream2].advance(uint8(v2))
+			br[stream].bitsRead += uint8(v)
+			br[stream].value <<= v & 63
+			br[stream2].bitsRead += uint8(v2)
+			br[stream2].value <<= v2 & 63
 		}
 
 		off += 4
@@ -1073,7 +1080,7 @@ func (d *Decoder) decompress4X8bit(dst, src []byte) ([]byte, error) {
 			}
 
 			// Read value and increment offset.
-			v := single[br.peekByteFast()>>shift].entry
+			v := single[uint8(br.value>>shift)].entry
 			nBits := uint8(v)
 			br.advance(nBits)
 			bitsLeft -= int(nBits)
@@ -1121,7 +1128,7 @@ func (d *Decoder) decompress4X8bitExactly(dst, src []byte) ([]byte, error) {
 	out := dst
 	dstEvery := (dstSize + 3) / 4
 
-	const shift = 0
+	const shift = 56
 	const tlSize = 1 << 8
 	const tlMask = tlSize - 1
 	single := d.dt.single[:tlSize]
@@ -1145,37 +1152,45 @@ func (d *Decoder) decompress4X8bitExactly(dst, src []byte) ([]byte, error) {
 			br[stream].fillFast()
 			br[stream2].fillFast()
 
-			v := single[br[stream].peekByteFast()>>shift].entry
+			v := single[uint8(br[stream].value>>shift)].entry
 			buf[off+bufoff*stream] = uint8(v >> 8)
-			br[stream].advance(uint8(v))
+			br[stream].bitsRead += uint8(v)
+			br[stream].value <<= v & 63
 
-			v2 := single[br[stream2].peekByteFast()>>shift].entry
+			v2 := single[uint8(br[stream2].value>>shift)].entry
 			buf[off+bufoff*stream2] = uint8(v2 >> 8)
-			br[stream2].advance(uint8(v2))
+			br[stream2].bitsRead += uint8(v2)
+			br[stream2].value <<= v2 & 63
 
-			v = single[br[stream].peekByteFast()>>shift].entry
+			v = single[uint8(br[stream].value>>shift)].entry
 			buf[off+bufoff*stream+1] = uint8(v >> 8)
-			br[stream].advance(uint8(v))
+			br[stream].bitsRead += uint8(v)
+			br[stream].value <<= v & 63
 
-			v2 = single[br[stream2].peekByteFast()>>shift].entry
+			v2 = single[uint8(br[stream2].value>>shift)].entry
 			buf[off+bufoff*stream2+1] = uint8(v2 >> 8)
-			br[stream2].advance(uint8(v2))
+			br[stream2].bitsRead += uint8(v2)
+			br[stream2].value <<= v2 & 63
 
-			v = single[br[stream].peekByteFast()>>shift].entry
+			v = single[uint8(br[stream].value>>shift)].entry
 			buf[off+bufoff*stream+2] = uint8(v >> 8)
-			br[stream].advance(uint8(v))
+			br[stream].bitsRead += uint8(v)
+			br[stream].value <<= v & 63
 
-			v2 = single[br[stream2].peekByteFast()>>shift].entry
+			v2 = single[uint8(br[stream2].value>>shift)].entry
 			buf[off+bufoff*stream2+2] = uint8(v2 >> 8)
-			br[stream2].advance(uint8(v2))
+			br[stream2].bitsRead += uint8(v2)
+			br[stream2].value <<= v2 & 63
 
-			v = single[br[stream].peekByteFast()>>shift].entry
+			v = single[uint8(br[stream].value>>shift)].entry
 			buf[off+bufoff*stream+3] = uint8(v >> 8)
-			br[stream].advance(uint8(v))
+			br[stream].bitsRead += uint8(v)
+			br[stream].value <<= v & 63
 
-			v2 = single[br[stream2].peekByteFast()>>shift].entry
+			v2 = single[uint8(br[stream2].value>>shift)].entry
 			buf[off+bufoff*stream2+3] = uint8(v2 >> 8)
-			br[stream2].advance(uint8(v2))
+			br[stream2].bitsRead += uint8(v2)
+			br[stream2].value <<= v2 & 63
 		}
 
 		{
@@ -1184,37 +1199,45 @@ func (d *Decoder) decompress4X8bitExactly(dst, src []byte) ([]byte, error) {
 			br[stream].fillFast()
 			br[stream2].fillFast()
 
-			v := single[br[stream].peekByteFast()>>shift].entry
+			v := single[uint8(br[stream].value>>shift)].entry
 			buf[off+bufoff*stream] = uint8(v >> 8)
-			br[stream].advance(uint8(v))
+			br[stream].bitsRead += uint8(v)
+			br[stream].value <<= v & 63
 
-			v2 := single[br[stream2].peekByteFast()>>shift].entry
+			v2 := single[uint8(br[stream2].value>>shift)].entry
 			buf[off+bufoff*stream2] = uint8(v2 >> 8)
-			br[stream2].advance(uint8(v2))
+			br[stream2].bitsRead += uint8(v2)
+			br[stream2].value <<= v2 & 63
 
-			v = single[br[stream].peekByteFast()>>shift].entry
+			v = single[uint8(br[stream].value>>shift)].entry
 			buf[off+bufoff*stream+1] = uint8(v >> 8)
-			br[stream].advance(uint8(v))
+			br[stream].bitsRead += uint8(v)
+			br[stream].value <<= v & 63
 
-			v2 = single[br[stream2].peekByteFast()>>shift].entry
+			v2 = single[uint8(br[stream2].value>>shift)].entry
 			buf[off+bufoff*stream2+1] = uint8(v2 >> 8)
-			br[stream2].advance(uint8(v2))
+			br[stream2].bitsRead += uint8(v2)
+			br[stream2].value <<= v2 & 63
 
-			v = single[br[stream].peekByteFast()>>shift].entry
+			v = single[uint8(br[stream].value>>shift)].entry
 			buf[off+bufoff*stream+2] = uint8(v >> 8)
-			br[stream].advance(uint8(v))
+			br[stream].bitsRead += uint8(v)
+			br[stream].value <<= v & 63
 
-			v2 = single[br[stream2].peekByteFast()>>shift].entry
+			v2 = single[uint8(br[stream2].value>>shift)].entry
 			buf[off+bufoff*stream2+2] = uint8(v2 >> 8)
-			br[stream2].advance(uint8(v2))
+			br[stream2].bitsRead += uint8(v2)
+			br[stream2].value <<= v2 & 63
 
-			v = single[br[stream].peekByteFast()>>shift].entry
+			v = single[uint8(br[stream].value>>shift)].entry
 			buf[off+bufoff*stream+3] = uint8(v >> 8)
-			br[stream].advance(uint8(v))
+			br[stream].bitsRead += uint8(v)
+			br[stream].value <<= v & 63
 
-			v2 = single[br[stream2].peekByteFast()>>shift].entry
+			v2 = single[uint8(br[stream2].value>>shift)].entry
 			buf[off+bufoff*stream2+3] = uint8(v2 >> 8)
-			br[stream2].advance(uint8(v2))
+			br[stream2].bitsRead += uint8(v2)
+			br[stream2].value <<= v2 & 63
 		}
 
 		off += 4
@@ -1280,7 +1303,7 @@ func (d *Decoder) decompress4X8bitExactly(dst, src []byte) ([]byte, error) {
 			}
 
 			// Read value and increment offset.
-			v := single[br.peekByteFast()>>shift].entry
+			v := single[br.peekByteFast()].entry
 			nBits := uint8(v)
 			br.advance(nBits)
 			bitsLeft -= int(nBits)