@klauspost 2021-09-07
Tons of unrelated deltas in benchmark. The related ones are better, though, so not sure what to think.